requirements.txt
pandas
numpy
scipy
matplotlib
seaborn
nibabel
network
nilearn
PyWavelets
mne
foof
statsmodels